fix(dom): solve race conditions around DOM initialization
This changes the internal workings of Cycle DOM in order to fix some race conditions that may occur
together with other drivers.

BREAKING CHANGE:
This is a breaking change that requires using also the latest Cycle Run version. May change
corner-case behavior of your app when it's initialized.

ISSUES CLOSED: #592